Renamed and compat packages to compatibility
