+++ common/src/main/java/org/adde0109/pcf/forwarding/compatibility/SpongePreLogin.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.forwarding.compat;
+package org.adde0109.pcf.forwarding.compatibility;
 
 import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
 import org.jetbrains.annotations.NotNull;
